+++ src/main/java/com/verdantartifice/thaumicwonders/core/ModMixins.java
@@ -1,5 +1,7 @@
 package com.verdantartifice.thaumicwonders.core;
 
+import com.verdantartifice.thaumicwonders.common.compat.ModIds;
+import zone.rong.mixinbooter.Context;
 import zone.rong.mixinbooter.ILateMixinLoader;
 
 import java.util.Collections;
@@ -9,5 +11,10 @@ public class ModMixins implements ILateMixinLoader {
     @Override
     public List<String> getMixinConfigs() {
         return Collections.singletonList("mixins.thaumicwonders.json");
+    }
+
+    @Override
+    public boolean shouldMixinConfigQueue(String mixinConfig) {
+        return !ModIds.tc4research.isLoaded && !ModIds.thaumtweaks.isLoaded;
     }
 }